add section table with data
